# perf(pnp): reduce work done at runtime