--- a/.pnp.js
+++ b/.pnp.js
@@ -36562,7 +36562,7 @@ var sources_path = __webpack_require__(9);
 class NodeFS extends FakeFS/* BasePortableFakeFS */.fS {
   constructor(realFs = (external_fs_default())) {
     super();
-    this.realFs = realFs; // @ts-ignore
+    this.realFs = realFs; // @ts-expect-error
 
     if (typeof this.realFs.lutimes !== `undefined`) {
       this.lutimesPromise = this.lutimesPromiseImpl;
@@ -36803,7 +36803,7 @@ class NodeFS extends FakeFS/* BasePortableFakeFS */.fS {
   }
 
   async lutimesPromiseImpl(p, atime, mtime) {
-    // @ts-ignore: Not yet in DefinitelyTyped
+    // @ts-expect-error: Not yet in DefinitelyTyped
     const lutimes = this.realFs.lutimes;
     if (typeof lutimes === `undefined`) throw ENOSYS(`unavailable Node binding`, `lutimes '${p}'`);
     return await new Promise((resolve, reject) => {
@@ -36812,7 +36812,7 @@ class NodeFS extends FakeFS/* BasePortableFakeFS */.fS {
   }
 
   lutimesSyncImpl(p, atime, mtime) {
-    // @ts-ignore: Not yet in DefinitelyTyped
+    // @ts-expect-error: Not yet in DefinitelyTyped
     const lutimesSync = this.realFs.lutimesSync;
     if (typeof lutimesSync === `undefined`) throw ENOSYS(`unavailable Node binding`, `lutimes '${p}'`);
     lutimesSync.call(this.realFs, sources_path/* npath.fromPortablePath */.cS.fromPortablePath(p), atime, mtime);
@@ -36921,12 +36921,12 @@ class NodeFS extends FakeFS/* BasePortableFakeFS */.fS {
   }
 
   watch(p, a, b) {
-    return this.realFs.watch(sources_path/* npath.fromPortablePath */.cS.fromPortablePath(p), // @ts-ignore
+    return this.realFs.watch(sources_path/* npath.fromPortablePath */.cS.fromPortablePath(p), // @ts-expect-error
     a, b);
   }
 
   watchFile(p, a, b) {
-    return this.realFs.watchFile(sources_path/* npath.fromPortablePath */.cS.fromPortablePath(p), // @ts-ignore
+    return this.realFs.watchFile(sources_path/* npath.fromPortablePath */.cS.fromPortablePath(p), // @ts-expect-error
     a, b);
   }
 
@@ -37195,12 +37195,12 @@ class ProxiedFS extends FakeFS/* FakeFS */.uY {
   }
 
   watch(p, a, b) {
-    return this.baseFs.watch(this.mapToBase(p), // @ts-ignore
+    return this.baseFs.watch(this.mapToBase(p), // @ts-expect-error
     a, b);
   }
 
   watchFile(p, a, b) {
-    return this.baseFs.watchFile(this.mapToBase(p), // @ts-ignore
+    return this.baseFs.watchFile(this.mapToBase(p), // @ts-expect-error
     a, b);
   }
 
@@ -37673,7 +37673,7 @@ function unwatchAllFiles(fakeFs) {
 const DEFAULT_COMPRESSION_LEVEL = `mixed`;
 
 function toUnixTimestamp(time) {
-  if (typeof time === `string` && String(+time) === time) return +time; // @ts-ignore
+  if (typeof time === `string` && String(+time) === time) return +time;
 
   if (Number.isFinite(time)) {
     if (time < 0) {
@@ -38352,15 +38352,23 @@ class ZipFS extends FakeFS/* BasePortableFakeFS */.fS {
         const data = Buffer.from(memory);
 
         if (compressionMethod === 0) {
+          this.fileSources.set(index, data);
           return data;
         } else if (opts.asyncDecompress) {
           return new Promise((resolve, reject) => {
             external_zlib_default().inflateRaw(data, (error, result) => {
-              error ? reject(error) : resolve(result);
+              if (error) {
+                reject(error);
+              } else {
+                this.fileSources.set(index, result);
+                resolve(result);
+              }
             });
           });
         } else {
-          return external_zlib_default().inflateRawSync(data);
+          const decompressedData = external_zlib_default().inflateRawSync(data);
+          this.fileSources.set(index, decompressedData);
+          return decompressedData;
         }
       } finally {
         this.libzip.fclose(file);
@@ -38644,7 +38652,7 @@ class ZipFS extends FakeFS/* BasePortableFakeFS */.fS {
 
   async readFilePromise(p, encoding) {
     // This is messed up regarding the TS signatures
-    if (typeof encoding === `object`) // @ts-ignore
+    if (typeof encoding === `object`) // @ts-expect-error
       encoding = encoding ? encoding.encoding : undefined;
     const data = await this.readFileBuffer(p, {
       asyncDecompress: true
@@ -38654,7 +38662,7 @@ class ZipFS extends FakeFS/* BasePortableFakeFS */.fS {
 
   readFileSync(p, encoding) {
     // This is messed up regarding the TS signatures
-    if (typeof encoding === `object`) // @ts-ignore
+    if (typeof encoding === `object`) // @ts-expect-error
       encoding = encoding ? encoding.encoding : undefined;
     const data = this.readFileBuffer(p);
     return encoding ? data.toString(encoding) : data;
@@ -39004,10 +39012,9 @@ class ZipOpenFS extends FakeFS/* BasePortableFakeFS */.fS {
     return await this.makeCallPromise(p, async () => {
       return await this.baseFs.realpathPromise(p);
     }, async (zipFs, {
-      archivePath,
       subPath
     }) => {
-      return this.pathUtils.join(await this.baseFs.realpathPromise(archivePath), this.pathUtils.relative(sources_path/* PortablePath.root */.LZ.root, await zipFs.realpathPromise(subPath)));
+      return this.pathUtils.join(zipFs.getRealPath(), this.pathUtils.relative(sources_path/* PortablePath.root */.LZ.root, await zipFs.realpathPromise(subPath)));
     });
   }
 
@@ -39015,10 +39022,9 @@ class ZipOpenFS extends FakeFS/* BasePortableFakeFS */.fS {
     return this.makeCallSync(p, () => {
       return this.baseFs.realpathSync(p);
     }, (zipFs, {
-      archivePath,
       subPath
     }) => {
-      return this.pathUtils.join(this.baseFs.realpathSync(archivePath), this.pathUtils.relative(sources_path/* PortablePath.root */.LZ.root, zipFs.realpathSync(subPath)));
+      return this.pathUtils.join(zipFs.getRealPath(), this.pathUtils.relative(sources_path/* PortablePath.root */.LZ.root, zipFs.realpathSync(subPath)));
     });
   }
 
@@ -39566,24 +39572,24 @@ class ZipOpenFS extends FakeFS/* BasePortableFakeFS */.fS {
 
   watch(p, a, b) {
     return this.makeCallSync(p, () => {
-      return this.baseFs.watch(p, // @ts-ignore
+      return this.baseFs.watch(p, // @ts-expect-error
       a, b);
     }, (zipFs, {
       subPath
     }) => {
-      return zipFs.watch(subPath, // @ts-ignore
+      return zipFs.watch(subPath, // @ts-expect-error
       a, b);
     });
   }
 
   watchFile(p, a, b) {
     return this.makeCallSync(p, () => {
-      return this.baseFs.watchFile(p, // @ts-ignore
+      return this.baseFs.watchFile(p, // @ts-expect-error
       a, b);
     }, (zipFs, {
       subPath
     }) => {
-      return zipFs.watchFile(subPath, // @ts-ignore
+      return zipFs.watchFile(subPath, // @ts-expect-error
       a, b);
     });
   }
@@ -40110,7 +40116,7 @@ function patchFs(patchedFs, fakeFs) {
       }
 
       setupFn(patchedFsPromises, `open`, async (...args) => {
-        // @ts-ignore
+        // @ts-expect-error
         const fd = await fakeFs.openPromise(...args);
         return new FileHandle(fd);
       }); // `fs.promises.realpath` doesn't have a `native` property
@@ -40323,7 +40329,7 @@ function getPathForDisplay(p) {
 
 
 function applyPatch(pnpapi, opts) {
-  // @ts-ignore
+  // @ts-expect-error
   const builtinModules = new Set(external_module_.Module.builtinModules || Object.keys(process.binding(`natives`)));
   /**
    * The cache that will be used for all accesses occuring outside of a PnP context.
@@ -40335,12 +40341,11 @@ function applyPatch(pnpapi, opts) {
    * a way to "reset" the environment temporarily)
    */
 
-  let enableNativeHooks = true; // @ts-ignore
+  let enableNativeHooks = true; // @ts-expect-error
 
-  process.versions.pnp = String(pnpapi.VERSIONS.std); // @ts-ignore
+  process.versions.pnp = String(pnpapi.VERSIONS.std);
 
-  const moduleExports = __webpack_require__(282); // @ts-ignore
-
+  const moduleExports = __webpack_require__(282);
 
   moduleExports.findPnpApi = lookupSource => {
     const lookupPath = lookupSource instanceof external_url_.URL ? (0,external_url_.fileURLToPath)(lookupSource) : lookupSource;
@@ -40402,15 +40407,14 @@ function applyPatch(pnpapi, opts) {
 
     const cacheEntry = entry.cache[modulePath];
     if (cacheEntry) return cacheEntry.exports; // Create a new module and store it into the cache
-    // @ts-ignore
+    // @ts-expect-error
 
-    const module = new external_module_.Module(modulePath, parent); // @ts-ignore
+    const module = new external_module_.Module(modulePath, parent); // @ts-expect-error
 
     module.pnpApiPath = moduleApiPath;
     entry.cache[modulePath] = module; // The main module is exposed as global variable
 
     if (isMain) {
-      // @ts-ignore
       process.mainModule = module;
       module.id = `.`;
     } // Try to load the module, and remove it from the cache if it fails
@@ -40419,7 +40423,7 @@ function applyPatch(pnpapi, opts) {
     let hasThrown = true;
 
     try {
-      // @ts-ignore
+      // @ts-expect-error
       module.load(modulePath);
       hasThrown = false;
     } finally {
@@ -40610,7 +40614,7 @@ function hydrateRuntimeState(data, {
   for (const [packageName, storeData] of data.packageRegistryData) {
     for (const [packageReference, packageInformationData] of storeData) {
       if (packageName === null !== (packageReference === null)) throw new Error(`Assertion failed: The name and reference should be null, or neither should`);
-      if (packageInformationData.discardFromLookup) continue; // @ts-ignore: TypeScript isn't smart enough to understand the type assertion
+      if (packageInformationData.discardFromLookup) continue; // @ts-expect-error: TypeScript isn't smart enough to understand the type assertion
 
       const packageLocator = {
         name: packageName,
@@ -40648,7 +40652,7 @@ function hydrateRuntimeState(data, {
 
 function makeApi(runtimeState, opts) {
   const alwaysWarnOnFallback = Number(process.env.PNP_ALWAYS_WARN_ON_FALLBACK) > 0;
-  const debugLevel = Number(process.env.PNP_DEBUG_LEVEL); // @ts-ignore
+  const debugLevel = Number(process.env.PNP_DEBUG_LEVEL); // @ts-expect-error
 
   const builtinModules = new Set(external_module_.Module.builtinModules || Object.keys(process.binding(`natives`))); // Splits a require request into its components, or return null if the request is a file path
 
@@ -40859,7 +40863,7 @@ function makeApi(runtimeState, opts) {
 
 
   function makeFakeModule(path) {
-    // @ts-ignore
+    // @ts-expect-error
     const fakeModule = new external_module_.Module(path, null);
     fakeModule.filename = path;
     fakeModule.paths = external_module_.Module._nodeModulePaths(path);
@@ -41431,19 +41435,23 @@ function makeManager(pnpapi, opts) {
   const apiMetadata = new Map([[initialApiPath, {
     cache: external_module_.Module._cache,
     instance: pnpapi,
-    stats: initialApiStats
+    stats: initialApiStats,
+    lastRefreshCheck: Date.now()
   }]]);
 
   function loadApiInstance(pnpApiPath) {
-    const nativePath = sources_path/* npath.fromPortablePath */.cS.fromPortablePath(pnpApiPath); // @ts-ignore
+    const nativePath = sources_path/* npath.fromPortablePath */.cS.fromPortablePath(pnpApiPath); // @ts-expect-error
 
-    const module = new external_module_.Module(nativePath, null); // @ts-ignore
+    const module = new external_module_.Module(nativePath, null); // @ts-expect-error
 
     module.load(nativePath);
     return module.exports;
   }
 
   function refreshApiEntry(pnpApiPath, apiEntry) {
+    const timeNow = Date.now();
+    if (timeNow - apiEntry.lastRefreshCheck < 500) return;
+    apiEntry.lastRefreshCheck = timeNow;
     const stats = opts.fakeFs.statSync(pnpApiPath);
 
     if (stats.mtime > apiEntry.stats.mtime) {
@@ -41464,7 +41472,8 @@ function makeManager(pnpapi, opts) {
       apiMetadata.set(pnpApiPath, apiEntry = {
         cache: {},
         instance: loadApiInstance(pnpApiPath),
-        stats: opts.fakeFs.statSync(pnpApiPath)
+        stats: opts.fakeFs.statSync(pnpApiPath),
+        lastRefreshCheck: Date.now()
       });
     }
 
@@ -41617,11 +41626,9 @@ if (__non_webpack_module__.parent && __non_webpack_module__.parent.id === `inter
   if (__non_webpack_module__.filename) {
     // We delete it from the cache in order to support the case where the CLI resolver is invoked from "yarn run"
     // It's annoying because it might cause some issues when the file is multiple times in NODE_OPTIONS, but it shouldn't happen anyway.
-    // @ts-ignore
     delete (external_module_default())._cache[__non_webpack_module__.filename];
   }
-} // @ts-ignore
-
+}
 
 if (process.mainModule === __non_webpack_module__) {
   const reportError = (code, message, data) => {