**What's the problem this PR addresses?**

The PnP runtime is doing a decent chunk of unnecessary work

**How did you fix it?**
- Cache decompression result in ZipFS
- Cache realPath result in ZipOpenFS
- throttle refreshApiEntry 

**Result**
When I tested this on https://github.com/yarnpkg/berry/issues/1669 just `perf(fslib): get real path from ZipFS in ZipOpenFS` alone reduces the build time from 2m28.568s to 1m34.816s (~54 seconds faster) (cc @alubbe)

Tested using 
```js
require('./.pnp.js').setup();
const { performance } = require('perf_hooks');

const start = performance.now();
for (let i = 0; i < 10000; i++) {
	require.resolve('typescript');
}
const end = performance.now();

console.log(end - start, 'ms');
```

||time|
|--|--|
|master| 7318ms|
|perf(fslib): cache decompression result|5235ms|
|perf(fslib): get real path from ZipFS in ZipOpenFS|2019ms|
|perf(pnp): throttle refreshApiEntry|1437ms|

**Checklist**
- [x] I have read the [Contributing Guide](https://yarnpkg.com/advanced/contributing).
- [x] I have set the packages that need to be released for my changes to be effective.
- [x] I have verified that all automated PR checks pass.